isAuthorized

+++ src/tarefas/tarefas.service.ts
@@ -2,20 +2,27 @@ import { ITarefaRepository } from './repository/itarefa.repository';
 import { TarefaCreateDto } from './dto/tarefa.create.dto';
 import { TarefaResponseDto } from './dto/tarefa.response.dto';
 import { ListaTarefaService } from '../lista_tarefa/lista_tarefa.service';
-import { Injectable} from '@nestjs/common';
+import { Injectable, UnauthorizedException} from '@nestjs/common';
 import { TarefaEntity } from './entity/tarefa.entity';
 import { TarefaUpdataDto } from './dto/tarefa.update.dto';
 import { BuscarTarefasQueryDto } from './dto/buscar_tarefas_query.dto';
 import { StatusFiltro } from 'src/common/enums/status_filtro.enum';
 import { Prisma } from 'generated/prisma/client';
 import { TarefaOrdenacao } from 'src/common/enums/tarefa_ordenacao.enum';
+import { ListaTarefaEntity } from 'src/lista_tarefa/entity/lista.tarefa.entity';
+import { IListaTarefaRepository } from 'src/lista_tarefa/repository/ilista.tarefas.repository';
 
 @Injectable()
 export class TarefasService {
-    constructor(private tarefaRepository: ITarefaRepository, private listaTarefaService: ListaTarefaService){}
+    constructor(
+        private tarefaRepository: ITarefaRepository,
+        private listaTarefaService: ListaTarefaService,
+        private listaTarefaRepository: IListaTarefaRepository    
+    ){}
 
     async create(idUserAuth: number, tarefaCreateDto: TarefaCreateDto): Promise<TarefaResponseDto>{
-        await this.listaTarefaService.retornePorId(tarefaCreateDto.idList, idUserAuth);
+        if(! await this.isAuthorized(tarefaCreateDto.idList, idUserAuth)) throw new UnauthorizedException();
+
         return await this.tarefaRepository.create(tarefaCreateDto);
     }
 
@@ -39,20 +46,23 @@ export class TarefasService {
     }
 
     async update(idUserAuth: number, idTarefa: number, tarefaUpdataDto: TarefaUpdataDto): Promise<TarefaResponseDto> {
-        await this.retorneTarefaAutorizada(idTarefa, idUserAuth);
+        const tarefa: TarefaEntity = await this.tarefaRepository.findById(idTarefa);
+        if(! await this.isAuthorized(tarefa.lista_id, idUserAuth)) throw new UnauthorizedException();
 
         const tarefaAtualizada = await this.tarefaRepository.update(idTarefa, tarefaUpdataDto);
         return this.toResponseDto(tarefaAtualizada);
     }
 
     async remove(idUserAuth: number, idTarefa: number): Promise<void> {
-        await this.retorneTarefaAutorizada(idTarefa, idUserAuth);
+        const tarefa: TarefaEntity = await this.tarefaRepository.findById(idTarefa);
+        if(! await this.isAuthorized(tarefa.lista_id, idUserAuth)) throw new UnauthorizedException();
 
         await this.tarefaRepository.delete(idTarefa);
     }
 
     private async retorneTarefaAutorizada(idTarefa: number, idUserAuth: number): Promise<TarefaEntity> {
-        const tarefa = await this.tarefaRepository.findById(idTarefa);
+        const tarefa: TarefaEntity = await this.tarefaRepository.findById(idTarefa);
+        if(! await this.isAuthorized(tarefa.lista_id, idUserAuth)) throw new UnauthorizedException();
 
         await this.listaTarefaService.retornePorId(tarefa.lista_id, idUserAuth);
 
@@ -105,5 +115,11 @@ export class TarefasService {
             vencido: tarefa.vencido
         };
     }
-    
+
+    private async isAuthorized(idLista: number, idUserAuth: number): Promise<boolean>{
+        const lista: ListaTarefaEntity = await this.listaTarefaRepository.findById(idLista);
+        if(lista.userId == idUserAuth) return true;
+
+        return false;
+    }
 }
